Cleaning up mock stat tests
