+++ tests/external_tests/stats/test_player_game_stats.py
@@ -13,7 +13,9 @@ class TestHittingStats(unittest.TestCase):
         cls.ty_france = 664034
         cls.shoei_game_id = 531368
         cls.ty_game_id = 715757
-                    
+        cls.cal_realeigh = 663728
+        cls.cal_game_id = 715757
+             
     @classmethod
     def tearDownClass(cls) -> None:
         pass
@@ -61,4 +63,34 @@ class TestHittingStats(unittest.TestCase):
         # game_stats should have vsplayer5y and playlog stats
         self.assertTrue(game_stats['hitting']['vsplayer5y'])
         self.assertTrue(game_stats['hitting']['playlog'])
+
+        splits = game_stats['hitting']['vsplayer5y']
+
+        for split in splits:
+            self.assertTrue(split.team)
+            self.assertTrue(split.stat)
+
+    def test_get_players_stats_for_cal_r(self):
+        """return player stat objects"""
+
+        game_stats = self.mlb.get_players_stats_for_game(person_id=self.cal_realeigh,
+                                                         game_id=self.cal_game_id)
+
+        # game stats should not be None
+        self.assertIsNotNone(game_stats)
+
+        # game_stats should be a dict
+        self.assertIsInstance(game_stats, dict)
+
+        # game_stats should have hitting stats
+        self.assertTrue(game_stats['hitting'])
+
+        # game_stats should have vsplayer5y and playlog stats
+        self.assertTrue(game_stats['hitting']['vsplayer5y'])
+        self.assertTrue(game_stats['hitting']['playlog'])
         
+        splits = game_stats['hitting']['vsplayer5y']
+
+        for split in splits:
+            self.assertTrue(split.team)
+            self.assertTrue(split.stat)
