+++ tests/mock_tests/stats/test_hitting_stats_mock.py
@@ -14,14 +14,20 @@ path_to_player_stats = os.path.join(current_directory, "../mock_json/stats/perso
 path_to_team_stats = os.path.join(current_directory, "../mock_json/stats/team/hitting_team_stats.json")
 path_to_not_found = os.path.join(current_directory, "../mock_json/response/not_found_404.json")
 path_to_error = os.path.join(current_directory, "../mock_json/response/error_500.json")
+path_to_hotcoldzone_file = os.path.join(current_directory, "../mock_json/stats/person/hotcoldzone.json")
+path_to_hitting_playlog_file = os.path.join(current_directory, "../mock_json/stats/person/hitting_player_playlog.json")
+path_to_hitting_pitchlog_file = os.path.join(current_directory, "../mock_json/stats/person/hitting_player_pitchlog.json")
 
+HOTCOLDZONE = open(path_to_hotcoldzone_file, "r", encoding="utf-8-sig").read()
 PLAYERSTATS = open(path_to_player_stats, "r", encoding="utf-8-sig").read()
 TEAMSTATS = open(path_to_team_stats, "r", encoding="utf-8-sig").read()
 NOT_FOUND_404 = open(path_to_not_found, "r", encoding="utf-8-sig").read()
 ERROR_500 = open(path_to_error, "r", encoding="utf-8-sig").read()
+HITTING_PLAY_LOG = open(path_to_hitting_playlog_file, "r", encoding="utf-8-sig").read()
+HITTING_PITCH_LOG = open(path_to_hitting_pitchlog_file, "r", encoding="utf-8-sig").read()
 
 @requests_mock.Mocker()
-class TestMlbDataApiMock(unittest.TestCase):
+class TestHittingStatsMock(unittest.TestCase):
     @classmethod
     def setUpClass(cls) -> None:
         cls.mlb = Mlb()
@@ -29,8 +35,11 @@ class TestMlbDataApiMock(unittest.TestCase):
         cls.team = cls.mlb.get_team(133)
         cls.mock_player_stats = json.loads(PLAYERSTATS)
         cls.mock_team_stats = json.loads(TEAMSTATS)
+        cls.mock_hotcoldzone = json.loads(HOTCOLDZONE)
         cls.error_500 = json.loads(ERROR_500)
         cls.mock_not_found = json.loads(NOT_FOUND_404)
+        cls.mock_hitting_playlog = json.loads(HITTING_PLAY_LOG)
+        cls.mock_hitting_pitchlog = json.loads(HITTING_PITCH_LOG)
 
     @classmethod
     def tearDownClass(cls) -> None:
@@ -108,3 +117,75 @@ class TestMlbDataApiMock(unittest.TestCase):
         self.assertTrue(season_advanced.season)
         self.assertTrue(season_advanced.stat.plateappearances)
         self.assertTrue(career_advanced.team)
+
+    def test_hitting_hotcoldzones_for_player(self, m):
+        """get_player_game_stats should return a dict with stats"""
+        m.get('https://statsapi.mlb.com/api/v1/people/665742/stats?stats=hotColdZones&group=hitting', json=self.mock_hotcoldzone,
+        status_code=200)
+        self.stats = ['hotColdZones']
+        self.groups = ['hitting']
+        stats = self.mlb.get_player_stats(self.player.id, stats=self.stats, groups=self.groups)
+
+        # game_stats should not be None
+        self.assertIsNotNone(stats)
+        
+        # game_stats should not be empty dic
+        self.assertNotEqual(stats, {})
+
+        self.assertTrue(stats['stats']['hotcoldzones'])
+
+        hotcoldzone = stats['stats']['hotcoldzones'][0]
+
+        # should not be empty
+        self.assertTrue(hotcoldzone.stat)
+
+        for zone in hotcoldzone.stat.zones:
+            self.assertTrue(zone.zone)
+
+    def test_hitting_pitchlog_for_player(self, m):
+        """get_player_game_stats should return a dict with stats"""
+        m.get('https://statsapi.mlb.com/api/v1/people/665742/stats?stats=pitchLog&group=hitting', json=self.mock_hitting_pitchlog,
+        status_code=200)
+        self.stats = ['pitchLog']
+        self.groups = ['hitting']
+        stats = self.mlb.get_player_stats(self.player.id, stats=self.stats, groups=self.groups)
+
+        # game_stats should not be None
+        self.assertIsNotNone(stats)
+        
+        # game_stats should not be empty dic
+        self.assertNotEqual(stats, {})
+
+        # playlog key should be populated
+        self.assertTrue('hitting' in stats)
+        self.assertTrue(stats['hitting']['pitchlog'])
+
+        # pitchlog items should have attr set
+        pitchlogs = stats['hitting']['pitchlog']
+
+        for pitchlog in pitchlogs:
+            self.assertTrue(pitchlog.stat)
+
+    def test_hitting_playlog_for_player(self, m):
+        """get_player_game_stats should return a dict with stats"""
+        m.get('https://statsapi.mlb.com/api/v1/people/665742/stats?stats=playLog&group=hitting', json=self.mock_hitting_playlog,
+        status_code=200)
+        self.stats = ['playLog']
+        self.groups = ['hitting']
+        stats = self.mlb.get_player_stats(self.player.id, stats=self.stats, groups=self.groups)
+
+        # game_stats should not be None
+        self.assertIsNotNone(stats)
+        
+        # game_stats should not be empty dic
+        self.assertNotEqual(stats, {})
+
+        # playlog key should be populated
+        self.assertTrue('hitting' in stats)
+        self.assertTrue(stats['hitting']['playlog'])
+
+        # pitchlog items should have attr set
+        playlogs = stats['hitting']['playlog']
+
+        for playlog in playlogs:
+            self.assertTrue(playlog.stat)
